[iOS] chore: TaskListDisplayLogic 파일 분리
- 뷰컨 안에 있으면 displayLogic을 테스트 할 때 뷰컨을 통째로 추가시켜야 해서 별도로 분리했습니다
